New r code for model 2
I changed the R code for model 2, since i had some problems in how to use python in R correctly. Now it should not be a problem. But somehow the individuals in my .tree file are converted to hermaphrodites during the filtering in R... So i will try to change the output from my model 2 in slim.